% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphawkes_functions.R
\name{br_plot}
\alias{br_plot}
\title{Plot of nonstationary background rate over space}
\usage{
br_plot(model, vals_include = FALSE)
}
\arguments{
\item{model}{the output from \code{misd()}}

\item{vals_include}{if TRUE, the background rate values will be printed on the map}
}
\value{
a ggplot of the background rates as a heat map
}
\description{
This function exports a heat map of the background rate of a nonstationary background rate over space, and is paired
with the \code{misd()} function when a nonstationary background rate is specified.
}
\examples{
data("hm.csv")
out = misd(dates = hm$t,
   ref_date = "1999-10-16",
   lat = hm$lat,
   lon = hm$lon,
   marks = hm$m,
   time_breaks = c(0,0.1, 0.5, 1,7,93,600),
   space_breaks = c(0,0.5, 1, 10, 25, 100),
   mark_breaks = c(3, 3.1,3.3, 4, 5, 8),
   just_times = TRUE,
   nonstat_br = TRUE)
br_p = br_plot(out, vals_include = TRUE)

}
