% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog-data.R
\docType{data}
\name{catalog}
\alias{catalog}
\title{Earthquake catalog from Jan 1, 1990 to June 9, 2016}
\format{An object of class list.}
\usage{
data(catalog)
}
\description{
Earthquake catalog from Jan 1, 1990 to June 9, 2016
}
\examples{
data(catalog)
with(catalog[catalog$Magnitude > 4.0,], plot(Longitude, Latitude))
}
\keyword{datasets}
