% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphawkes_functions.R
\name{ci_hist}
\alias{ci_hist}
\title{Histogram of super-thinned process}
\usage{
ci_hist(superthin, nbins = 30)
}
\arguments{
\item{superthin}{the output from \code{super_thin()}}

\item{nbins}{scalar of the number of bins for the histogram}
}
\value{
a histogram of the super-thinned process.
}
\description{
This function exports a histogram of the super-thinned process.
}
\examples{
data("hm.csv")
out = misd(dates = hm$t,
   ref_date = "1999-10-16",
   lat = hm$lat,
   lon = hm$lon,
   marks = hm$m,
   time_breaks = c(0,0.1, 0.5, 1,7,93,600),
   space_breaks = c(0,0.5, 1, 10, 25, 100),
   mark_breaks = c(3, 3.1,3.3, 4, 5, 8),
   just_times = T)
st = super_thin(out)
ci_h = ci_hist(st)
}
