% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphawkes_functions.R
\name{ci_plot}
\alias{ci_plot}
\title{Plot of Conditional Intensity Over Time}
\usage{
ci_plot(
  model,
  min_date = model$input$ref_date,
  max_date = model$data$dates[nrow(model$data)],
  superthin
)
}
\arguments{
\item{model}{the output from \code{misd()}}

\item{min_date}{the minimum date to be shown on the plot}

\item{max_date}{the maximum date to be shown on the plot}

\item{superthin}{the output from \code{super_thin()}}
}
\value{
a plot that shows the estimated conditional intensity as a black line with the number of monthly events as
gray vertical lines
}
\description{
In this function, the estimated conditional intensity is plotted against
the number of events. The estimated conditional intensity is found by taking the
median value in each month and multiplying it by the number of days in that month.
}
\examples{
data("hm.csv")
out = misd(dates = hm$t,
   ref_date = "1999-10-16",
   lat = hm$lat,
   lon = hm$lon,
   marks = hm$m,
   time_breaks = c(0,0.1, 0.5, 1,7,93,600),
   space_breaks = c(0,0.5, 1, 10, 25, 100),
   mark_breaks = c(3, 3.1,3.3, 4, 5, 8),
   just_times = T)
st = super_thin(out)
ci_p = ci_plot(out, st)

}
