% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeDistance}
\alias{ComputeDistance}
\title{Compute the distance between events.}
\usage{
ComputeDistance(x, y, n)
}
\arguments{
\item{x}{A vector of length n, holding the longitudes of interest}

\item{y}{A vector of length n, holding the latitudes of interest}

\item{n}{A single integer indicating the number of events}
}
\value{
dist A vector of length n(n-1)/2, holding the distance between events
}
\description{
Compute the distance between events.
}
