% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeTimeDifference}
\alias{ComputeTimeDifference}
\title{Compute the time differences between events.}
\usage{
ComputeTimeDifference(x)
}
\arguments{
\item{x}{A vector of length n, holding the times of interest}
}
\value{
tdiff A vector of length n(n-1)/2, holding the time differences
}
\description{
Compute the time differences between events.
}
