% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeTriggeringMarkedSpatialTemporalNH}
\alias{ComputeTriggeringMarkedSpatialTemporalNH}
\title{Computes the triggering function for a marked spatial temporal hawkes point process with a inhomonogenous and stationary background rate.}
\usage{
ComputeTriggeringMarkedSpatialTemporalNH(g, f, kp, mu, n)
}
\arguments{
\item{g}{A vector of length n(n-1)/2, holding the triggering probabilities for time}

\item{f}{A vector of length n(n-1)/2, holding the triggering probabilities for distance}

\item{kp}{A vector of length n(n-1)/2, holding the triggering probabilities for mark productivity}

\item{mu}{A vector of length n, holding the background rate in the grid of each event}

\item{n}{A single integer indicating the number of events in the catalog.}
}
\value{
pbNew A vector of background probabilities

ptNew A vector of triggering probabilities
}
\description{
Computes the triggering function for a marked spatial temporal hawkes point process with a inhomonogenous and stationary background rate.
}
