% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeTriggeringTemporal}
\alias{ComputeTriggeringTemporal}
\title{Initialize triggering and background probabilities}
\usage{
ComputeTriggeringTemporal(g, mu, n)
}
\arguments{
\item{g}{a vector of length n(n-1)/2, holding the triggering probabilities}

\item{mu}{a numeric of length 1, holding the background rate}

\item{n}{A single integer indicating the number of events in the catalog.}
}
\value{
pbNew a vector of background probabilities

ptNew a vector of triggering probabilities
}
\description{
Initialize triggering and background probabilities
}
