% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeTriggeringTemporalNS}
\alias{ComputeTriggeringTemporalNS}
\title{Computes the triggering function for a spatial temporal hawkes point process with a homonogenous and stationary background rate.}
\usage{
ComputeTriggeringTemporalNS(g, kp, mu, n)
}
\arguments{
\item{g}{A vector of length n(n-1)/2, holding the triggering probabilities for time}

\item{kp}{A single number representing the mean number of events triggered per day}

\item{mu}{A vector of length n(n-1)/2, holding the background probabilities associated with the grid of each point}

\item{n}{A single integer indicating the number of events in the catalog.}
}
\description{
Computes the triggering function for a spatial temporal hawkes point process with a homonogenous and stationary background rate.
}
