% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distanceFrom}
\alias{distanceFrom}
\title{Computes the distance from one events to all other events and sort them is ascending order}
\usage{
distanceFrom(x1, y1, j)
}
\arguments{
\item{x1}{A vector of length n, holding the longitudes of interest}

\item{y1}{A vector of length n, holding the latitude of interest}

\item{j}{A single integer indicating the position of the origin event}
}
\value{
dist1 A sorted vector of distances from the origin event
}
\description{
Computes the distance from one events to all other events and sort them is ascending order
}
