% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distanceFrom2}
\alias{distanceFrom2}
\title{Computes the distance from one events to all other events and sort them is ascending order}
\usage{
distanceFrom2(x0, y0, x1, y1)
}
\arguments{
\item{x0}{A single double indicating the position of the origin event}

\item{y0}{A single double indicating the position of the origin event}

\item{x1}{A vector of length n, holding the longitudes of interest}

\item{y1}{A vector of length n, holding the latitude of interest}
}
\value{
dist A vector of distances from the origin event
}
\description{
Computes the distance from one events to all other events and sort them is ascending order
}
