% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETASsim-data.R
\docType{data}
\name{ETASsim}
\alias{ETASsim}
\title{One simulation of ETAS}
\format{An object of class list.}
\usage{
data(ETASsim)
}
\description{
One simulation of ETAS
}
\examples{
data(ETASsim)
times <- ETASsim$t
}
\keyword{datasets}
