% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDataPrep.R
\name{fitDataPrep}
\alias{fitDataPrep}
\title{Prepares nphawkes model estimation and nphawkes data for nphawkes forecasting}
\usage{
fitDataPrep(fit, data)
}
\arguments{
\item{fit}{An object of class nphawkesMSTNH, nphawkesMSTNA, or nphawkesMSTNAC}

\item{data}{An object of class nphData}
}
\description{
Prepares nphawkes model estimation and nphawkes data for nphawkes forecasting
}
