% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastDataPrep.R
\name{forecastDataPrep}
\alias{forecastDataPrep}
\title{Prepares nphawkes model estimation and nphawkes data for nphawkes forecasting}
\usage{
forecastDataPrep(fit, data, forecast_date, catalog_start = "1985-01-01")
}
\arguments{
\item{fit}{An object of class nphawkesMSTNH, nphawkesMSTNA, or nphawkesMSTNAC}

\item{data}{An object of class nphData}

\item{forecast_date}{The date for which to generate a daily forecast}

\item{catalog_start}{The first date of the catalog using in model fitting used to calculate the tot_time parameter used in the thinning of background events}
}
\value{
data A list of outputs list(est_mu, nbins_mu, x_window, y_window, Nb, xmin, ymin, tot_time, t_obs, x_obs, y_obs, m_obs)
}
\description{
Prepares nphawkes model estimation and nphawkes data for nphawkes forecasting
}
\examples{
data(catalog)
data <- nphData(data = catalog[catalog$Magnitude > 4.0,], time_var = 'tdiff', x_var = 'Longitude', y_var = 'Latitude', mag = 'Magnitude')
fit <- nphawkesMSTNH(data = data, np = 100, eps = 1e-2, xrange = range(data$lon), yrange = range(data$lat))
myForecastData <- forecastDataPrep(fit = fit, data = data, forecast_date = as.Date(max(data$DateTime)) + 1, catalog_start = min(data$DateTime))
str(myForecastData)
}
