% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBackgroudPosition}
\alias{getBackgroudPosition}
\title{Computes the grid locations of each event for the background rate}
\usage{
getBackgroudPosition(nbins_x, nbins_y, x, y, grid_x, grid_y, n)
}
\arguments{
\item{nbins_x}{A single number holding the number of longitudes bins for the background rate}

\item{nbins_y}{A single number holding the number of latitude bins for the background rate}

\item{x}{A vector of length n, holding the longitudes of interest}

\item{y}{A vector of length n, holding the latitude of interest}

\item{grid_x}{A vector holding the grid positions for the longitudes bins of the background rate}

\item{grid_y}{A vector holding the grid positions for the latitude bins of the background rate}

\item{n}{A single integer indicating the number of events in the catalog}
}
\value{
D A vector of length nbins_x  * nbins_y holding vectors with event indices if they fall within the grid cell
}
\description{
Computes the grid locations of each event for the background rate
}
