% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBackground}
\alias{getBackground}
\title{Computes the background probability associated with each point in the catalog}
\usage{
getBackground(x, y, t, pbNew, nn, nbins_x, nbins_y, grid_x, grid_y, delta_x,
  delta_y, n)
}
\arguments{
\item{x}{A vector of length n, holding the longitudes of interest}

\item{y}{A vector of length n, holding the latitude of interest}

\item{t}{A Single number, holding the maximum time of interest}

\item{pbNew}{A vector of length n, holding the background probabilities}

\item{nn}{An integer for the n-th nearest neighbor to use in background smoothing}

\item{nbins_x}{A single number holding the number of longitudes bins for the background rate}

\item{nbins_y}{A single number holding the number of latitude bins for the background rate}

\item{grid_x}{A vector holding the grid positions for the longitudes bins of the background rate}

\item{grid_y}{A vector holding the grid positions for the latitude bins of the background rate}

\item{delta_x}{A single number holding the bin width longitudes bins for the background rate}

\item{delta_y}{A single number holding the bin height of latitude bins for the background rate}

\item{n}{A single integer indicating the number of events in the catalog}
}
\value{
mu A vector of length n holding the new background probabilities
}
\description{
Computes the background probability associated with each point in the catalog
}
