% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getDistanceBins}
\alias{getDistanceBins}
\title{Determines the distance bins for use in the triggering function}
\usage{
getDistanceBins(nbins_r, dist, grid_r)
}
\arguments{
\item{nbins_r}{The number of distance bins on which to estimate the spatial triggering of the Hawkes Point Process Model}

\item{dist}{A vector of length n(n-1)/2 holding the distance between events}

\item{grid_r}{The grid on which the spatial triggering distribution is estimated}
}
\description{
Determines the distance bins for use in the triggering function
}
