% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMagnitudeBins}
\alias{getMagnitudeBins}
\title{Determines the magnitude bins for use in the triggering function}
\usage{
getMagnitudeBins(nbins_m, mags, grid_m)
}
\arguments{
\item{nbins_m}{The number of distance bins on which to estimate the magnitude productivity of the Hawkes Point Process Model}

\item{mags}{A vector of length n(n-1)/2 holding the mainshock magnitudes between events}

\item{grid_m}{The grid on which the magnitude productivity is estimated}
}
\description{
Determines the magnitude bins for use in the triggering function
}
