% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMags}
\alias{getMags}
\title{Generates marks for all sequences of events in the catalog}
\usage{
getMags(m, n)
}
\arguments{
\item{m}{A vector of length n, holding the marks}

\item{n}{A single integer indicating the number of events}
}
\value{
mags A vector of length n(n-1)/2 containing the marks associated with each triggering probability
}
\description{
Generates marks for all sequences of events in the catalog
}
