% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getNN}
\alias{getNN}
\title{Determines the distance of the specified nearest neighbor}
\usage{
getNN(x, y, np, n)
}
\arguments{
\item{x}{A vector of length n, holding the longitudes of interest}

\item{y}{A vector of length n, holding the latitude of interest}

\item{np}{A single integer indicating the nth nearest neighbor}

\item{n}{A single integer indicating the  number of events}
}
\value{
nn A vector of nearest neighbor distances
}
\description{
Determines the distance of the specified nearest neighbor
}
