% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getTimeBins}
\alias{getTimeBins}
\title{Determines the time bins for use in the triggering function}
\usage{
getTimeBins(nbins_t, time, grid_t)
}
\arguments{
\item{nbins_t}{The number of time bins on which to estimate the temporal triggering of the Hawkes Point Process Model}

\item{time}{A vector of length n(n-1)/2 holding the time between events}

\item{grid_t}{The grid on which the temporal triggering distribution is estimated}
}
\description{
Determines the time bins for use in the triggering function
}
