% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HawkesSim1-data.R
\docType{data}
\name{HawkesSim1}
\alias{HawkesSim1}
\title{One simulation of a Hawkes Process}
\format{An object of class list.}
\usage{
data(HawkesSim1)
}
\description{
One simulation of a Hawkes Process
}
\examples{
data(HawkesSim1)
plot(x = HawkesSim1$t.bkgrd, y = 1:HawkesSim1$n1, xlab = 'time', ylab = 'event number')
}
\keyword{datasets}
