% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm}
\alias{hm}
\title{Hector Mine earthquake catalog.}
\format{
A data frame with 542 rows and 4 variables:
\describe{
  \item{lon}{longitude}
  \item{lat}{latitude}
  \item{t}{time since the initial earthquake event, measured in days}
  \item{m}{magnitude}
}
}
\source{
Hector mine data courtesy of the U.S. Geological Survey
}
\usage{
hm
}
\description{
A dataset containing the location, magnitude, and time since initial earthquake
event occurring on  October 16, 1999.
}
\keyword{datasets}
