% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initProb}
\alias{initProb}
\title{Initialize triggering and background probabilities}
\usage{
initProb(n)
}
\arguments{
\item{n}{A single integer indicating the number of events in the catalog.}
}
\description{
Initialize triggering and background probabilities
}
