% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_shooting_functions.R
\name{nph}
\alias{nph}
\title{Model Independent Stochastic Declustering}
\usage{
nph(
  dates,
  ref_date = min(dates),
  lat = rep(0, length(dates)),
  lon = rep(0, length(dates)),
  marks = rep(0, length(dates)),
  g_bins = c(0, 1),
  h_bins = c(0, 1),
  k_bins = c(0, 1),
  g_quantile = FALSE,
  k_quantile = FALSE,
  h_quantile = FALSE,
  k_length = 6,
  h_length = 6,
  g_length = 6,
  stopwhen = 0.001,
  time_unit = "day",
  dist_unit = "mile"
)
}
\arguments{
\item{dates}{a vector of dates, as LIST DATE FORMATS ALLOWED}

\item{ref_date}{a date to serve as time 0, defaults to earliest observation}

\item{lat}{a vector of latitudes, omit if not using spatial data}

\item{lon}{a vector of longitudes, omit if not using spatial data}

\item{marks}{a vecotr of marks, or magnitudes, omit if not using marked process}

\item{g_bins}{a vector of cutoff values for temporal bins of time differences}

\item{h_bins}{a vector of cutoff values for spatial bins of distance differences}

\item{k_bins}{a vector of cutoff values for magnitude bins}

\item{g_quantile}{FALSE by default to use defined temporal bins, TRUE to establish unifrom bins}

\item{k_quantile}{FALSE by default to use defined magnitude bins, TRUE to establish unifrom bins}

\item{h_quantile}{FALSE by default to use defined spatial bins, TRUE to establish uniform bins}

\item{k_length}{a scalar to define the number of desired magnitude bins}

\item{h_length}{a scalar to define the number of desired unifrom spatial bins CHANGE THIS}

\item{g_length}{a scalar to define the number of desired uniform temporal bins CHANGE THIS IN CODE}

\item{time_unit}{character string that specifies the desired unit of time}

\item{dist_unit}{character string that specifies the desired unit of distance: meter, kilometer, or mile}

\item{stop_when}{scalar that serves as conversion criterion, 1e-3 as default}
}
\value{
Probability matrix \code{p0} containing the probabilities that event
\code{i} is an offspring of event \code{j}, \code{i > j}. Diagonal elements
represent the probability that event \code{i} is a background event.

\code{g} is a vector of the estimated values for each bin of the temporal triggering component

\code{h} is a vector of the estimated values for each bin of the spatial triggering component

\code{k} is a vector of the estimataed values for each bin of the magnitude triggering component

\code{br} is the estimated background rate of the process

\code{perc_diag} is the proportion of mass lying on the diagonal of matrix \code{p0}

\code{time_bins} is a matrix containing the temporal bin of each pair of events
 @return \code{dist_bins} is a matrix containing the spatial bin of each pair of events
 @return \code{mark_bins} is a vector containing the magnitude bin of each event
 @return \code{input} is a list of all inputs
}
\description{
This function uses nonparametric procedures to analyze a Hawkes
process in temporal or spatio-temporal domain, with or without marks.
}
