% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphawkesMSTNH.R
\name{nphawkesMSTNH}
\alias{nphawkesMSTNH}
\title{Fits a Marked Spatial Temporal Nonparametric Hawkes Point Process Model with a nonhomogenous and stationary background rate}
\usage{
nphawkesMSTNH(data, nbins_m = 8, nbins_t = 25, nbins_r = 25, np = 50,
  num_iter = 1000, nbins_x = 100, nbins_y = 100, xrange = NULL,
  yrange = NULL, eps = 0.001, verbose = T, warn = T)
}
\arguments{
\item{data}{A list or data frame with the times, locations, and magnitudes of interest}

\item{nbins_m}{The number of bins on which to estimate the magnitude productivity function for Hawkes Point Process Model}

\item{nbins_t}{The number of bins on which to estimate the spatial triggering of the Hawkes Point Process Model}

\item{nbins_r}{The number of bins on which to estimate the temporal triggering of the Hawkes Point Process Model}

\item{np}{The nearest neighbor used for bandwidth selection}

\item{num_iter}{The maximum number of iterations to run}

\item{nbins_x}{The number of x bins on which to estimate the inhomogenous background rate}

\item{nbins_y}{The number of y bins on which to estimate the inhomogenous background rate}

\item{xrange}{The x range on which estimate the spatial triggering function}

\item{yrange}{The y range on which estimate the spatial triggering function}

\item{eps}{The convergence criteria}

\item{verbose}{A logicial to control the amount of information printed to the console during estimation}

\item{warn}{A logical to control the size of the dataset}
}
\value{
out An object containing the input data and fitted model
}
\description{
Fits a Marked Spatial Temporal Nonparametric Hawkes Point Process Model with a nonhomogenous and stationary background rate
}
\examples{
data(ETASsim)
newdata <- nphData(data = ETASsim, time_var = 't', x_var = 'lon', y_var = 'lat', mag = 'hm')
res <- nphawkesMSTNH(data = newdata, xrange = c(0,4), yrange = c(0,6), eps = 1e-2)

}
