% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphawkesT.R
\name{nphawkesT}
\alias{nphawkesT}
\title{Fits a Temporal Nonparametric Hawkes Point Process Model with a stationary background rate}
\usage{
nphawkesT(data, nbins_t = 25, num_iter = 1000, eps = 0.001, verbose = T,
  warn = T)
}
\arguments{
\item{data}{A list or data frame with the times of interest}

\item{nbins_t}{The number of bins on which to estimate the  Hawkes Point Process Model}

\item{num_iter}{The maximum number of iterations to run}

\item{eps}{The convergence criteria}

\item{verbose}{A logicial to control the amount of information printed to the console during estimation}

\item{warn}{A logical to control the size of the dataset}
}
\value{
out An object containing the input data and fitted model
}
\description{
Fits a Temporal Nonparametric Hawkes Point Process Model with a stationary background rate
}
\examples{
data(ETASsim)
newdata <- nphData(data = ETASsim, time_var = 't')
res <- nphawkesT(data = newdata, nbins_t = 30)
}
