% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphawkesTNS.R
\name{nphawkesTNS}
\alias{nphawkesTNS}
\title{Fits a Temporal Nonparametric Hawkes Point Process Model with a non stationary background rate}
\usage{
nphawkesTNS(data, nbins_t = 25, nbins_mu = 10, bw = 100,
  num_iter = 1000, eps = 0.001, verbose = T, warn = T)
}
\arguments{
\item{data}{A list or data frame with the times of interest}

\item{nbins_t}{The number of bins on which to estimate the  Hawkes Point Process Model}

\item{nbins_mu}{The number of bins on which to estimate the background rate of the Hawkes Point Process Model}

\item{bw}{The background smoothing bandwidth}

\item{num_iter}{The maximum number of iterations to run}

\item{eps}{The convergence criteria}

\item{verbose}{A logicial to control the amount of information printed to the console during estimation}

\item{warn}{A logical to control the size of the dataset}
}
\value{
out An object containing the input data and fitted model
}
\description{
Fits a Temporal Nonparametric Hawkes Point Process Model with a non stationary background rate
}
\examples{
data(HawkesSim1)
newdata <- nphData(data = HawkesSim1, time_var = 't', x_var = 'lon', y_var = 'lat',mag = 'hm')
fit <- nphawkesTNS(data = newdata, nbins_t = 100, nbins_mu = 50)
str(fit)
}
