% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphData.R
\name{nphData}
\alias{nphData}
\title{Prepares data for nphawkes estimation}
\usage{
nphData(data = NULL, time_var = NULL, x_var = NULL, y_var = NULL,
  mag = NULL)
}
\arguments{
\item{data}{A list or data frame with event information}

\item{time_var}{A character string representing the name of the time variable in the current data set}

\item{x_var}{A character string representing the name of the x variable (lon) in the current data set}

\item{y_var}{A character string representing the name of the y variable (lat) in the current data set}

\item{mag}{A character string represting the name of the magnitude variable in the current data set}
}
\value{
data A dataset with modified names
}
\description{
Prepares data for nphawkes estimation
}
\examples{
mydata <- data.frame(times = runif(100), lon = runif(100, 0, 4), lat = runif(100, 0, 6), magnitudes = rexp(100))
newdata <- nphData(data = mydata, time_var = 'times', x_var = 'lon', y_var = 'lat', mag = 'magnitudes')
res <- nphawkesMSTNH(data = newdata)
}
