% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pKern}
\alias{pKern}
\title{Calculates a probability weighted normal kernel}
\usage{
pKern(prob, x, y, bwd)
}
\arguments{
\item{prob}{A single value on which to calculate the kernel}

\item{x}{A single number holding the longitude of interest}

\item{y}{A single number holding the latitude of interest}

\item{bwd}{A single number holding the bandwith used in the kernel}
}
\value{
value A single value of the kernel density at the point of interest
}
\description{
Calculates a probability weighted normal kernel
}
