% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pKernT}
\alias{pKernT}
\title{Kernel smoothing for nonstationary temporal background rate}
\usage{
pKernT(prob, t, bwd)
}
\arguments{
\item{prob}{A single value on which to calculate the kernel}

\item{t}{A single number holding the time of interest}

\item{bwd}{A single number holding the bandwith used in the kernel}
}
\description{
Kernel smoothing for nonstationary temporal background rate
}
