% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nphawkesT.R
\name{plot.nphawkesT}
\alias{plot.nphawkesT}
\title{Function to plot the magnitude productivity, spatial, temporal, and background components of the Hawkes model}
\usage{
\method{plot}{nphawkesT}(x, print = FALSE, ...)
}
\arguments{
\item{x}{An object of class nphawkesMSTH}

\item{print}{A logical indicating whether the plot should be printed or returned}

\item{...}{Other parameters passed in}

\item{type}{A string indicating the type of plot to generate}
}
\value{
p A ggplot2 plot
}
\description{
Function to plot the magnitude productivity, spatial, temporal, and background components of the Hawkes model
}
\examples{
data(catalog)
data <- nphData(data = catalog[catalog$Magnitude > 4.0,], time_var = 'tdiff', x_var = 'Longitude', y_var = 'Latitude', mag = 'Magnitude')
fit <- nphawkesT(data = data)
plot(x = fit, type = 'time')
}
