% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nphawkesTNS.R
\name{plot.nphawkesTNS}
\alias{plot.nphawkesTNS}
\title{Function to plot the magnitude productivity, spatial, temporal, and background components of the Hawkes model}
\usage{
\method{plot}{nphawkesTNS}(x, type = "time", smooth = FALSE,
  print = FALSE, ...)
}
\arguments{
\item{x}{An object of class nphawkesMSTH}

\item{type}{A string indicating the type of plot to generate}

\item{print}{A logical indicating whether the plot should be printed or returned}

\item{...}{Other parameters passed in}
}
\value{
p A ggplot2 plot
}
\description{
Function to plot the magnitude productivity, spatial, temporal, and background components of the Hawkes model
}
\examples{
data(HawkesSim1)
newdata <- nphData(data = HawkesSim1, time_var = 't', x_var = 'lon', y_var = 'lat',mag = 'hm')
fit <- nphawkesTNS(data = newdata, nbins_t = 100, nbins_mu = 50)
plot(x = fit, type = 'time')
plot(x = fit, type = 'background')
plot(x = fit, type = 'background', smooth = TRUE)

}
