% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{replaceMu}
\alias{replaceMu}
\title{Utility function to replace elements in vector mu with indices specified in vector of vectors D}
\usage{
replaceMu(mu, D, n)
}
\arguments{
\item{mu}{A vector of length nbins_x  * nbins_y holding the background rate in each cell}

\item{D}{A vector of length nbins_x  * nbins_y holding vectors with event indices if they fall within the grid cell}

\item{n}{A single integer indicating the number of events in the catalog}
}
\value{
muvec A vector length n, holding the updated probabilities
}
\description{
Utility function to replace elements in vector mu with indices specified in vector of vectors D
}
