% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_shooting_functions.R
\name{se_bars}
\alias{se_bars}
\title{Standard Error Estimation}
\usage{
se_bars(model)
}
\arguments{
\item{model}{the output from \code{nph()}}
}
\value{
a list of variance estimates for all utilized triggering component
}
\description{
This function estimates the variance of each bin for all utilized triggering components.
}
