% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphawkes_functions.R
\name{st_plot}
\alias{st_plot}
\title{Super-thinning Plot}
\usage{
st_plot(superthin, method = "superthin", time_label = "Year")
}
\arguments{
\item{superthin}{the output from \code{super_thin()}}

\item{method}{chacter string denoting if the residual analysis method utilized was
"superthin", "thin", or "superpose"}

\item{time_label}{character string that provides the frequency of tick marks on the x-axis}
}
\value{
a tiered plot with superposed points on the top tier, points that were retained, not thinned,
on the middle tier, and points that were thinned on the bottom tier.
}
\description{
This function exports a plot displaying the performance of the super-thinning procedure.
}
\examples{
data("hm.csv")
out = nph(dates = hm$t,
   ref_date = "1999-10-16",
   lat = hm$lat,
   lon = hm$lon,
   marks = hm$m,
   time_breaks = c(0,0.1, 0.5, 1,7,93,600),
   space_breaks = c(0,0.5, 1, 10, 25, 100),
   mark_breaks = c(3, 3.1,3.3, 4, 5, 8),
   just_times = T)
st = super_thin(out)
sp = st_plot(st, "superthin")
}
