% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumProbs}
\alias{sumProbs}
\title{Utility function to check that the rows of the matrix for the background and triggering probabilities sum to 1}
\usage{
sumProbs(vt, vb, sz)
}
\arguments{
\item{vt}{A vector of length n(n-1)2, holding the triggering probabilities}

\item{vb}{A vector of length n, holding the background probabilities}

\item{sz}{A single integer indicating the number of events}
}
\description{
Utility function to check that the rows of the matrix for the background and triggering probabilities sum to 1
}
