% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphawkes_functions.R
\name{super_thin}
\alias{super_thin}
\title{Super-thinning}
\usage{
super_thin(
  K = "median_ci",
  model,
  method = "superthin",
  map = world,
  region = ".",
  sim_grid = TRUE,
  lat_lim = model$input$lat_lim,
  lon_lim = model$input$lon_lim
)
}
\arguments{
\item{K}{a constant or character string that governs the amount of thinning and
superposing that is implemented. Can be a constant value, the median, mean, minimum, or maximum
 conditional intensity: "median_ci", "mean_ci", "min_ci", or "max_ci", respectively.}

\item{model}{the output from \code{misd()}}

\item{method}{character string that defines residual analysis method as "superthin", "thin", or "superpose"}

\item{map}{name of map provided by the maps package, defaults to maps::world}

\item{region}{name of subregion to include, defaults to entirety of map}

\item{sim_grid}{TRUE if geographical coordinates do not necessarily pertain to
certain area, such as 0x1 by 0x1 grid, FALSE if using specific region or not using spatial information}

\item{lat_lim}{vector containing minimum and maximum latitude bounds if sim_grid is TRUE}

\item{lon_lim}{vector containing minimum and maximum longitude bounds if sim_gird is TRUE}
}
\value{
a data frame which includes the time, location, and estimated conditional intensity of events.
The type of event, either observed or simulated, is noted along with the probability that the event was kept
and whether or not the point was in fact retained.
}
\description{
This function performs super-thinning to assess model fit. Thinning only can be executed by
simply considering events of type "thin" or "retain", while superpositioning only can be executed
by simply treating "thinned" points as "retained".
}
\details{
This function is to be used in conjunction with the \code{misd()} function from the \code{nphawkes} library.

To simulate spatial data, the user may define the \code{map} and \code{region} to easily simulate points within
set political borders. This feature is not quite implementable, but may be soon.
Otherwise, simulated points may be established when \code{sim_grid = TRUE}, based on minimum and maximum observed
latitude and longitude coordinates.
}
\examples{
data("hm.csv")
out = misd(dates = hm$t,
   ref_date = "1999-10-16",
   lat = hm$lat,
   lon = hm$lon,
   marks = hm$m,
   time_breaks = c(0,0.1, 0.5, 1,7,93,600),
   space_breaks = c(0,0.5, 1, 10, 25, 100),
   mark_breaks = c(3, 3.1,3.3, 4, 5, 8),
   just_times = T)
st = super_thin(K = "min_ci",
    model = out,
    method = "superthin",
    )

}
