% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_shooting_functions.R
\name{trig_plots}
\alias{trig_plots}
\title{Triggering Plots}
\usage{
trig_plots(
  model,
  g_max = max(model$g_bins),
  k_max = max(model$k_bins),
  k_min = min(model$k_bins),
  h_max = max(model$h_bins),
  mag_label = "magnitude"
)
}
\arguments{
\item{model}{the output from \code{nph()}}

\item{g_max}{the maximum x axis value shown in the temporal plot}

\item{k_max}{the maximum x axis value shown in the magnitude plot}

\item{k_min}{the minimum x axis value shown in the magnitude plot}

\item{h_max}{the maximum x axis value shown in the spatial plot}

\item{mag_label}{character string representing what the magnitude measures}
}
\value{
histogram estimaotrs for all utilized triggering components
}
\description{
This function exports histogram estimators for all utilized triggering components. Plots show
estimated value of triggering functions for each bin along with \pm 2 standard errors bars
}
