% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_shooting_functions.R
\name{cond_int}
\alias{cond_int}
\title{Conditional Intensity}
\usage{
cond_int(model)
}
\arguments{
\item{model}{the output from \code{nph()}}
}
\value{
a data frame containing the time, location, marks, and estimated conditional intensity
}
\description{
This function estimates the conditional intensity function of the observed process.
}
