% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphawkes_functions.R
\name{st_plot}
\alias{st_plot}
\title{Super-thinning Plot}
\usage{
st_plot(
  superthin,
  method = "superthin",
  time_label = "Year",
  plot_title = NULL
)
}
\arguments{
\item{superthin}{the output from \code{super_thin()}}

\item{method}{chacter string denoting if the residual analysis method utilized was
"superthin", "thin", or "superpose"}

\item{time_label}{character string that provides the frequency of tick marks on the x-axis}

\item{plot_title}{character string for the plot title}
}
\value{
a tiered plot with superposed points on the top tier, points that were retained, not thinned,
on the middle tier, and points that were thinned on the bottom tier.
}
\description{
This function exports a plot displaying the performance of the super-thinning procedure.
}
