% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_shooting_functions.R
\name{st_plot}
\alias{st_plot}
\title{Super-thinning Plot}
\usage{
st_plot(superthin)
}
\arguments{
\item{superthin}{the output from \code{super_thin()}}
}
\value{
a tiered plot with superposed points on the top tier, points that were retained, not thinned,
on the middle tier, and points that were thinned on the bottom tier.
}
\description{
This function exports a plot displaying the performance of the super-thinning procedure.
}
