% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_shooting_functions.R
\name{super_thin}
\alias{super_thin}
\title{Super-thinning}
\usage{
super_thin(
  K = "median_ci",
  model,
  method = "superthin",
  map = world,
  region = "."
)
}
\arguments{
\item{K}{a constant or character string that governs the amount of thinning and
superposing that is implemented. Can be a constant value, the median, mean, minimum, or maximum
 conditional intensity: "median_ci", "mean_ci", "min_ci", or "max_ci", respectively.}

\item{model}{the output from \code{nph()}}

\item{method}{character string that defines residual analysis method as "superthin", "thin", or "superpose"}

\item{map}{name of map provided by the maps package, defaults to maps::world}

\item{region}{name of subregion to include, defaults to entirety of map}
}
\value{
a data frame which includes the time, location, and estimted conditional intensity of events.
The type of event, either observed or simulated, is noted along with the probability that the event was kept
and wheter or not the point was in fact retained.
}
\description{
This function performs super-thinning to assess model fit. Thinning only can be executed by
simply considering events of type "thin" or "retain", while superpositioning only can be executed
by simply treating "thinned" points as "retained".
}
