% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_shooting_functions.R
\name{trig_plots}
\alias{trig_plots}
\title{Triggering Plots}
\usage{
trig_plots(
  model,
  g_xlim = c(min(model$time_breaks), max(model$time_breaks)),
  h_xlim = c(min(model$space_breaks), max(model$space_breaks)),
  k_xlim = c(min(model$mark_breaks), max(model$mark_breaks)),
  g_ylim = c(0, NA),
  h_ylim = c(0, NA),
  k_ylim = c(0, NA),
  mag_label = "magnitude",
  plot_title = NULL
)
}
\arguments{
\item{model}{the output from \code{nph()}}

\item{g_xlim}{vector of minimum and maximum x-axis value shown in the temporal plot}

\item{h_xlim}{vector of minimum and maximum x-axis value shown in the spatial plot}

\item{k_xlim}{vector of minimum and maximum x-axis value shown in the magnitude plot}

\item{g_ylim}{vector of minimum and maximum y-axis value shown in the temporal plot}

\item{h_ylim}{vector of minimum and maximum y-axis value shown in the spatial plot}

\item{k_ylim}{vector of minimum and maximum y-axis value shown in the magnitude plot}

\item{mag_label}{character string representing what the magnitude measures}

\item{title}{character string for the title of the collection of triggering plots}
}
\value{
histogram estimators for all utilized triggering components
}
\description{
This function exports histogram estimators for all utilized triggering components. Plots show
estimated value of triggering functions for each bin along with \pm 2 standard errors bars
}
