% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphawkes_functions.R
\name{trig_plots}
\alias{trig_plots}
\title{Triggering Plots}
\usage{
trig_plots(
  model,
  time_xlim = c(min(model$time_breaks), max(model$time_breaks)),
  space_xlim = c(min(model$space_breaks), max(model$space_breaks)),
  mark_xlim = c(min(model$mark_breaks), max(model$mark_breaks)),
  time_ticks = NULL,
  space_ticks = NULL,
  mark_ticks = NULL,
  time_ylim = c(0, NA),
  space_ylim = c(0, NA),
  mark_ylim = c(0, NA),
  mag_label = "magnitude",
  se_include = TRUE
)
}
\arguments{
\item{model}{the output from \code{misd()}}

\item{time_xlim}{vector of minimum and maximum x-axis value shown in the temporal plot}

\item{space_xlim}{vector of minimum and maximum x-axis value shown in the spatial plot}

\item{mark_xlim}{vector of minimum and maximum x-axis value shown in the magnitude plot}

\item{time_ylim}{vector of minimum and maximum y-axis value shown in the temporal plot}

\item{space_ylim}{vector of minimum and maximum y-axis value shown in the spatial plot}

\item{mark_ylim}{vector of minimum and maximum y-axis value shown in the magnitude plot}

\item{mag_label}{character string representing what the magnitude measures}

\item{se_include}{TRUE to include standard error estimates for bins of triggering components, FALSE
to exclude them from the visualizations}
}
\value{
\code{all_plots} is a cowplot object of histogram estimators for all utilized triggering components

\code{time_plot} is a ggplot object of the histogram estimator for the temporal effect

\code{space_plot} is a ggplot object of the histogram estimator for the spatial effect,
NULL if data provided is not a spatial process

\code{mark_plot} is a ggplot object of the histogram estimator for the mark effect,
NULL if data provided is not a marked process
}
\description{
This function exports histogram estimators for all utilized triggering components. Plots show
estimated value of triggering functions for each bin along with \eqn{+/- 2} standard errors bars
}
\details{
Depending on bin size, x-axis labels may overlap, impairing readability. Axis tick marks may be changed for plots
using \code{scale_x_continuous()} within the \code{ggplot2} library.
}
\examples{
data("hm.csv")
out = misd(dates = hm$t,
   ref_date = "1999-10-16",
   lat = hm$lat,
   lon = hm$lon,
   marks = hm$m,
   time_breaks = c(0,0.1, 0.5, 1,7,93,600),
   space_breaks = c(0,0.5, 1, 10, 25, 100),
   mark_breaks = c(3, 3.1,3.3, 4, 5, 8),
   just_times = T)
tp = trig_plots(out,
    time_xlim = c(0, 2),
    space_xlim = c(0, 10),
    mark_xlim = c(3.9, 5.1))

}
